+++ models/save.js
@@ -52,6 +52,7 @@ module.exports = (sequelize, DataTypes) => {
     {
       sequelize,
       modelName: "Save",
+      timestamps: false
     }
   );
   return Save;
