+++ routes/posts.routes.js
@@ -16,7 +16,8 @@ postsRouter.get('/',authMiddleware, postsController.getPosts);
 //게시글상세조회
 postsRouter.get('/:postId',authMiddleware, postsController.getOnePosts);
 //게시글 삭제
-postsRouter.delete('/:postId',authMiddleware, imagesController.deleteImage, postsController.deletePosts);
+//postsRouter.delete('/:postId',authMiddleware, imagesController.deleteImage);
+postsRouter.delete('/:postId',authMiddleware, postsController.deletePosts);
 
 //게시글 저장 (찜하기)
 postsRouter.put("/:postId", authMiddleware, postsController.savePosts);
